DATAES-659 - Move MappingElasticsearchConverter to correct package.
Original PR: #328
