+++ src/main/java/org/springframework/data/elasticsearch/core/index/MappingParameters.java
@@ -12,7 +12,8 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */package org.springframework.data.elasticsearch.core.indexmapping;
+ */
+package org.springframework.data.elasticsearch.core.index;
 
 import java.io.IOException;
 import java.lang.annotation.Annotation;
