+++ src/main/java/org/springframework/data/elasticsearch/core/index/package-info.java
@@ -2,4 +2,4 @@
  * infrastructure to define the Elasticsearch mapping for an index.
  */
 @org.springframework.lang.NonNullApi
-package org.springframework.data.elasticsearch.core.indexmapping;
+package org.springframework.data.elasticsearch.core.index;
